% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DnaX_mesh}
\alias{DnaX_mesh}
\title{DnaX and FtsZ timelapse of D39 S.pneumoniae. Cells are tracked using Oufti and converted to "mesh" dataframe using BactMAP::extr_Oufti}
\format{A data frame with columns:
\describe{
 \item{X}{x-coordinate of cell outline point (pixels)}
 \item{Y}{y-coordinate of cell outline point (pixels)}
 \item{cell}{cell identifier. unique per frame.}
 \item{frame}{frame identifier. in this case, an image was taken every 20 seconds, so 1 frame corresponds to 20s.}
 \item{num}{number indicating the order of cell outline points (useful for plotting polygons)}
 \item{max.width}{maximum cell width in pixels}
 \item{length}{the point of cell length from pole 0}
 \item{steplength}{the distance (over the length axis) of this cell outline point from the previous cell outline point}
 \item{max.length}{the maximum cell length of the given cell}
 \item{xy}{indicates whether it's the right or the left side of the cell (inherited from oufti)}
 \item{area}{cell area in pixels(2)}
 \item{angle}{the angle of the cells length axis to the horizontal line of the image.}
 \item{Xmid}{the mid-point (pixel x-coordinate) of the cell}
 \item{Ymid}{the mid-point (pixel y-coordinate) of the cell}
 \item{X_rot}{the x-coordinate of the cell outline point (in pixels) when the cell length axis is horizontal, with mid-cell at (0,0)}
 \item{Y_rot}{the y-coordinate of the cell outline point (in pixels) when the cell length axis is horizontal, with mid-cell at (0,0)}
 \item{max_um}{the maximum cell length in micron}
 \item{maxwum}{the maximum cell width in micron}
 \item{Xrotum}{the x-coordinate of the cell outline point (in micron) when the cell length axis is horizontal, with mid-cell at (0,0)}
 \item{Yrotum}{the y-coordinate of the cell outline point (in micron) when the cell lenth axis is horizontal, with mid-cell at (0,0)}
}}
\source{
van Raaphorst, Renske, Morten Kjos, and Jan-Willem Veening. "Chromosome segregation drives division site selection in Streptococcus pneumoniae." Proceedings of the National Academy of Sciences 114.29 (2017): E5959-E5968.
}
\usage{
DnaX_mesh
}
\description{
DnaX and FtsZ timelapse of D39 S.pneumoniae. Cells are tracked using Oufti and converted to "mesh" dataframe using BactMAP::extr_Oufti
}
\examples{
\dontrun{
 DnaX_mesh
}
}
\keyword{datasets}
