% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DnaX_tracks}
\alias{DnaX_tracks}
\title{DnaX and FtsZ timelapse of D39 S.pneumoniae. DnaX spots are tracked using ISbatch and converted using extr_ISBatch}
\format{A data frame with columns:
\describe{
 \item{x}{x-coordinate of detected DnaX spot}
 \item{Y}{y-coordinate of detected DnaX spot}
 \item{cell}{cell identifier. unique per frame.}
 \item{frame}{frame identifier. in this case, an image was taken every 20 seconds, so 1 frame corresponds to 20s.}
 \item{max.width}{maximum cell width in pixels}
 \item{max.length}{the maximum cell length of the given cell}
 \item{l}{relative localization of DnaX on the length axis of the cell.}
 \item{d}{relative localization of DnaX on the width axis (diameter) of the cell}
 \item{trajectory}{identifier of the DnaX track}
 \item{trajectory_length}{the time (in frames) the trajectory could be followed}
 \item{displacement_sq}{the displacement of the molecule (in pixels, as calculated by ISBatch)}
}}
\source{
van Raaphorst, Renske, Morten Kjos, and Jan-Willem Veening. "Chromosome segregation drives division site selection in Streptococcus pneumoniae." Proceedings of the National Academy of Sciences 114.29 (2017): E5959-E5968.
}
\usage{
DnaX_tracks
}
\description{
DnaX and FtsZ timelapse of D39 S.pneumoniae. DnaX spots are tracked using ISbatch and converted using extr_ISBatch
}
\examples{
\dontrun{
 DnaX_tracks
}
}
\keyword{datasets}
