% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_Ilastik.R
\name{getFilesIlastik}
\alias{getFilesIlastik}
\title{Function to take Ilastik cell information and cell masks}
\usage{
getFilesIlastik(
  directory,
  name_classes = c("attack", "inside"),
  smoothing = TRUE,
  mag = "No_PixelCorrection"
)
}
\description{
To be able to use this function, you need one directory with the Object Predictions (saved as TIFF) and Object Tables (saved as CSV).
Save them with the names the software gives them ("*_Object Predictions.TIFF" and "_table.csv").
}
