% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{onePerCell}
\alias{onePerCell}
\title{Function to get a dataframe per cell out of your mesh, spots, object or tiff dataset.

\code{onePerCell()} returns a dataframe with one table per cell, simplified especially for people who wish
to make their own plots (e.g. histograms, etc) based on the summarized cell data. Automatically, it groups
the input dataset by unique cell (cell & frame), but it is possible to use this function with other column names.}
\usage{
onePerCell(inputframe, by = c("cell", "frame"))
}
\arguments{
\item{inputframe}{the dataset (in form of a dataframe) to be summarized}

\item{by}{the grouping value. default = c("cell", "frame")}
}
\value{
the function will return a dataframe (tibble) with the data summarized by the grouping value \code{by}.
it will do this by returning the common values per cell (mean intensity, etc), removing values that are useless
in one-column context (x/y coordinates of outlines), and calculating the mean, median and standard deviation of
numerical values it doesn't recognize. if there are character values with different values per grouped variable,
the function will take ONLY THE FIRST VARIABLE.
}
\description{
Function to get a dataframe per cell out of your mesh, spots, object or tiff dataset.

\code{onePerCell()} returns a dataframe with one table per cell, simplified especially for people who wish
to make their own plots (e.g. histograms, etc) based on the summarized cell data. Automatically, it groups
the input dataset by unique cell (cell & frame), but it is possible to use this function with other column names.
}
