% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{orientCells}
\alias{orientCells}
\title{Orient your cells by the side in which your spots or objects (predominantly) are located

\code{orientCells()} takes a \code{spots_relative} or \code{object_relative} dataset and uses the relative localization of each spot or mid-point of each object
to flip all related cells & content to one side based on this localization.}
\usage{
orientCells(dataTurn, cellData, removeNonOriented = FALSE)
}
\arguments{
\item{dataTurn}{has to be a \code{spots_relative} or \code{object_relative} dataset}

\item{cellData}{a list of corresponding datasets that need to be turned as well. can be 1 dataset (for instance \code{mesh}) or a list of datasets (like \code{list(mesh, object_relative)})}
}
\value{
the function will return a list of the datasets, where the relative X/Y coordinates of the cells and internal features are all oriented to one side.
}
\description{
Orient your cells by the side in which your spots or objects (predominantly) are located

\code{orientCells()} takes a \code{spots_relative} or \code{object_relative} dataset and uses the relative localization of each spot or mid-point of each object
to flip all related cells & content to one side based on this localization.
}
\examples{
\dontrun{

output <- orientCells(dataTurn = myData$spots_relative, cellData = list(myData$mesh, myData$object_relative))
}

}
