\name{plotRaw}
\alias{plotRaw}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot TIFF, cell contours and/or spot localization on the original x/y space.
}
\description{
To check segmentation, view TIFF images and/or show the localization of spots inside cells. ggplot2-based, so other geoms and theme layers can be added.
}
\usage{
plotRaw(tiffdata,
        meshdata,
        spotdata,
        frameN = 1,
        xrange,
        yrange,
        viridisoption = "inferno",
        meshcolor = "white",
        spotcolor = "yellow",
        valuerange)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tiffdata}{
(list of) dataframe(s) containing *x* and *y* coordinates of pixels and their *values*. Can be generated by loading a TIFF into R using bactMAP::extr_OriginalStack()
}
  \item{meshdata}{
dataframe containing (at least) *X* and *Y* coordinates of cells outlines, the *frame* number of the image and *cell* numbers.
}
  \item{spotdata}{
dataframe containing (at least) *x* and *y* coordinates of spots and the *frame* number of the image.
}
  \item{frameN}{
number of the frame to be displayed. default = 1.
}
  \item{xrange}{
vector with minimum and maximum x coordinates to be displayed (by default all is shown).
}
  \item{yrange}{
vector with minimum and maximum y coordinates to be displayed (by default all is shown).
}
  \item{viridisoption}{
one of the viridis color scale options (default = "inferno") for the displayed TIFF. options: "magma", "viridis", "cividis", "plasma" and "inferno"
}
  \item{meshcolor}{
color of the cell outlines. default = "white"
}
  \item{spotcolor}{
color for displayed spots. default = "yellow"
}
  \item{valuerange}{
vector with minumum and maximum pixel values to show TIFF, use this option to increase contrast
}
}

\value{

A plot showing either a TIFF, mesh outlines, spot coordinates or a combination of the three.

}

\author{
Renske van Raaphorst
}


\seealso{
\code{\link{extr_Original}}, \href{https://cran.r-project.org/web/packages/viridis/vignettes/intro-to-viridis.html}{viridis}}

\examples{
\dontrun{

#load TIF image from BactMAP examples. go to veeninglab.com/bactmap to download.

TIF_image <- file.choose()

#use extr_-function to turn into list of dataframes

TIF_df <- bactMAP::extr_OriginalStack(TIF_image)

#plot image
plotRaw(TIF_image)

require(ggplot2)
#zoom in and add title
plotRaw(TIF_image, xrange=(c(200,500)), yrange=c(400,600)) + ggplot2::ggtitle("This is a TIF image")
}
}
