\name{bactMAP}
\alias{bactMAP-package}
\alias{bactMAP}
\alias{BACTMAP}

\docType{package}

\title{Analyze Bacterial Cell Segmentation And Fluoresence Data

}
\description{Package to upload an uniformally analyze (bacterial) cell segmentation and fluorescence data, making it possible to combine output from different analysis tools into one or more datasets. Vizualize the data using ggplot2, statistically compare datasets and automatically produce a summary of the data.

}

\references{van Raaphorst R*, Kjos M*, Veening JW (2017) Chromosome segregation drives division site selection in \emph{Streptococcus pneumoniae. Proc Natl Acad Sci U S A.} 114(29):E5959-E5968.

}
\author{
Renske van Raaphorst<renske.vanraaphorst@unil.ch>
University of Lausanne
}
\note{
Compatible software outputs:

*MicrobeJ<\url{www.microbej.com}>
*ObjectJ<\url{https://sils.fnwi.uva.nl/bcb/objectj/}>
*ISBatch<\url{http://singlemolecule.github.io/iSBatch/}>
*Utrack<\url{http://www.utsouthwestern.edu/labs/danuser/software/#utrack_anc}>
*Oufti<\url{www.oufti.org}>
*MicrobeTracker<\url{www.microbetracker.org}>
*Morphometrics<\url{https://simtk.org/projects/morphometrics}>

To be added:
*DeepCell<\url{https://covertlab.github.io/DeepCell/about/}>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~


\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

