% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TurnedCell4}
\alias{TurnedCell4}
\title{DnaX and FtsZ timelapse of D39 S.pneumoniae. This dataframe contains the fluorescence and mesh information of the FtsZ channel of cell number 4. The data is retreived by uploading the TIFF of FtsZ-RFP using extr_OriginalStacks and subsequently running extr_OriginalCells with this and the dataset bactMAP::DnaX_mesh. For disk space reasons, only cell 4 is saved here. Get the whole dataset on veeninglab.com/bactmap.}
\format{A data frame with columns:
\describe{
 \item{x}{x-coordinate of the pixel}
 \item{y}{y-coordinate of the pixel}
 \item{cell}{cell identifier. unique per frame.}
 \item{frame}{frame identifier. in this case, an image was taken every 20 seconds, so 1 frame corresponds to 20s.}
 \item{max.width}{maximum cell width in pixels}
 \item{max.length}{the maximum cell length of the given cell}
 \item{pointN}{number of specific pixel. unique per cell and frame}
 \item{values}{intensity of the pixel}
 \item{X_rot}{pixel x coordinate when the cell is turned so that the length axis is horizontal and mid-cell is at 0,0.}
 \item{Y_rot}{pixel y coordinate when the cell is turned so that the length axis is horizontal and mid-cell is at 0,0.}
 \item{xt}{the point marking one corner of the square pixel (x-coordinate)}
 \item{yt}{the point marking one corner of the square pixel (y-coordinate)}
 \item{area}{the cell area (pixel squared)}
}}
\source{
van Raaphorst, Renske, Morten Kjos, and Jan-Willem Veening. "Chromosome segregation drives division site selection in Streptococcus pneumoniae." Proceedings of the National Academy of Sciences 114.29 (2017): E5959-E5968.
}
\usage{
TurnedCell4
}
\description{
DnaX and FtsZ timelapse of D39 S.pneumoniae. This dataframe contains the fluorescence and mesh information of the FtsZ channel of cell number 4. The data is retreived by uploading the TIFF of FtsZ-RFP using extr_OriginalStacks and subsequently running extr_OriginalCells with this and the dataset bactMAP::DnaX_mesh. For disk space reasons, only cell 4 is saved here. Get the whole dataset on veeninglab.com/bactmap.
}
\examples{
\dontrun{
 TurnedCell4
}

}
\keyword{datasets}
