% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VanFL_meshWT}
\alias{VanFL_meshWT}
\title{VanFL dataset: cell outlines (mesh) WT}
\format{a "mesh" data frame (\code{\link{mesh}}) describing the shape and size of Wild Type S. pneumoniae cells.}
\source{
van Raaphorst, Renske, Morten Kjos, and Jan-Willem Veening. "Chromosome segregation drives division site selection in Streptococcus pneumoniae." Proceedings of the National Academy of Sciences 114.29 (2017): E5959-E5968
}
\usage{
VanFL_meshWT
}
\description{
\describe{Cells were stained with fluorescent vancomycin and imaged as described in \href{https://www.pnas.org/content/early/2017/06/30/1620608114}{van Raaphorst, Kjos & Veening, 2017}. The cells and fluorescent objects where segmented using Oufti and imported in R using BactMAP's extr_Oufti function.}
}
\examples{
\dontrun{
 View(VanFL_meshWT)
 }

}
\keyword{datasets}
