\name{bactmapnews}
\alias{bactmapnews}

\title{This is version 0.1.0.2 of BactMAP}

\description{
This version contains a few bug-fixes when it comes to importing data from MicrobeJ and Oufti. Check the wiki and/or documentation for more information on the import functions! It also contains two new functions for analysis after import: orientCells and onePerCell. These functions can be useful to flip cells based on spot or object localization (orientCells) and to summarize data when you're not very familiar with R (onePerCell). Check out the help (by typing ?orientCells or ?onePerCell) to find out how they work.
}


